+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Unreleased
 
-* Changed
-  * Testing section docs updated to use `jest.spyOn` instead of `jest.mock`.
+* Fixed
+  * Updated `terra-native-select` example to fix slightly truncated button.
 
+* Changed
+  * Updated `terra-button-group` tests.
+  * Updated testing recommendations to use `jest.spyOn` instead of `jest.mock`.
+  
 ## 1.38.0 - (August 23, 2023)
 
 * Removed
@@ -71,6 +75,7 @@
 
 * Changed
   * Updated Toggle Button Examples to have a dashed border around interactive elements.
+
 ## 1.31.0 - (July 13, 2023)
 
 * Added